refactor: separate resource code generation
