+++ codegen/lib/connect.ts
@@ -6,8 +6,8 @@ import {
   type EndpointsLayoutContext,
   setEndpointsLayoutContext,
 } from './layouts/endpoints.js'
+import { getResourceLayoutContexts } from './layouts/resources.js'
 import {
-  renderBlueprintResources,
   type RouteIndexLayoutContext,
   type RouteLayoutContext,
   setRouteLayoutContext,
@@ -93,7 +93,7 @@ export const connect = (
   routeIndexes['']?.add('seam-http-endpoints-without-workspace.js')
 
   const resourceExports: string[] = []
-  for (const resource of renderBlueprintResources(blueprint)) {
+  for (const resource of getResourceLayoutContexts(blueprint)) {
     files[`${resourcesPath}/${resource.fileName}`] = {
       contents: Buffer.from(
         `/* Automatically generated by codegen/smith.ts. */\n\nexport type ${resource.typeName} = ${resource.type}\n`,
