+++ codegen/lib/layouts/route.ts
@@ -1,14 +1,8 @@
-import type {
-  Blueprint,
-  Endpoint,
-  Namespace,
-  Parameter,
-  Property,
-  Resource,
-  Route,
-} from '@seamapi/blueprint'
+import type { Endpoint, Namespace, Parameter, Route } from '@seamapi/blueprint'
 import type { Method } from 'axios'
 import { camelCase, kebabCase, pascalCase } from 'change-case'
+
+import { getResourceTypeName } from './resources.js'
 
 export interface RouteLayoutContext {
   className: string
@@ -183,43 +177,6 @@ const renderResponse = (endpoint: Endpoint): string => {
   return `{ ${JSON.stringify(responseKey)}: ${value} }`
 }
 
-const renderResource = (resource: Resource): string =>
-  renderObject(resource.properties, renderProperty)
-
-export const getResourceTypeName = (resourceType: string): string =>
-  `${pascalCase(resourceType)}Resource`
-
-export const renderBlueprintResources = (
-  blueprint: Blueprint,
-): Array<{ fileName: string; typeName: string; type: string }> => {
-  const resources = [
-    ...blueprint.resources,
-    ...blueprint.events,
-    ...blueprint.actionAttempts,
-  ]
-  const resourceTypes = [
-    ...new Set(resources.map(({ resourceType }) => resourceType)),
-  ]
-  return ['unknown', ...resourceTypes].map((resourceType) => {
-    if (resourceType === 'unknown') {
-      return {
-        fileName: 'unknown.ts',
-        typeName: 'UnknownResource',
-        type: 'Record<string, unknown>',
-      }
-    }
-    const type = resources
-      .filter((resource) => resource.resourceType === resourceType)
-      .map(renderResource)
-      .join(' | ')
-    return {
-      fileName: `${kebabCase(resourceType)}.ts`,
-      typeName: getResourceTypeName(resourceType),
-      type: type || 'Record<string, unknown>',
-    }
-  })
-}
-
 const renderObject = <T extends { name: string }>(
   members: T[],
   render: (member: T) => string,
@@ -249,26 +206,6 @@ const renderParameter = (parameter: Parameter): string => {
     return `Array<${renderScalar(parameter.itemFormat, parameter)}>`
   }
   return renderScalar(parameter.format, parameter)
-}
-
-const renderProperty = (property: Property): string => {
-  let type: string
-  if (property.format === 'object') {
-    type = renderObject(property.properties, renderProperty)
-  } else if (property.format === 'list') {
-    if (property.itemFormat === 'object') {
-      type = `Array<${renderObject(property.itemProperties, renderProperty)}>`
-    } else if (property.itemFormat === 'discriminated_object') {
-      type = `Array<${property.variants
-        .map(({ properties }) => renderObject(properties, renderProperty))
-        .join(' | ')}>`
-    } else {
-      type = `Array<${renderScalar(property.itemFormat, property)}>`
-    }
-  } else {
-    type = renderScalar(property.format, property)
-  }
-  return property.isNullable ? `${type} | null` : type
 }
 
 const renderScalar = (format: string, value: unknown): string => {
